Break out a types crate which can be compiled by SP1
