+++ tide-disco/src/request.rs
@@ -5,55 +5,15 @@
 // along with the tide-disco library. If not, see <https://mit-license.org/>.
 
 use crate::method::Method;
-use serde::{Deserialize, Serialize};
-use snafu::{OptionExt, Snafu};
+use snafu::OptionExt;
 use std::any::type_name;
 use std::collections::HashMap;
 use std::fmt::Display;
-use strum_macros::EnumString;
 use tagged_base64::TaggedBase64;
 use tide::http::{self, Headers, content::Accept, mime::Mime};
 use vbs::{BinarySerializer, Serializer, version::StaticVersionType};
 
-#[derive(Clone, Debug, Snafu, Deserialize, Serialize)]
-pub enum RequestError {
-    #[snafu(display("missing required parameter: {}", name))]
-    MissingParam { name: String },
-
-    #[snafu(display(
-        "incorrect parameter type: {} cannot be converted to {}",
-        actual,
-        expected
-    ))]
-    IncorrectParamType {
-        actual: RequestParamType,
-        expected: RequestParamType,
-    },
-
-    #[snafu(display("value {} is too large for type {}", value, expected))]
-    IntegerOverflow { value: u128, expected: String },
-
-    #[snafu(display("Unable to deserialize from JSON"))]
-    Json,
-
-    #[snafu(display("Unable to deserialize from binary"))]
-    Binary,
-
-    #[snafu(display("Unable to deserialise from tagged base 64: {}", reason))]
-    TaggedBase64 { reason: String },
-
-    #[snafu(display("Content type not specified or type not supported"))]
-    UnsupportedContentType,
-
-    #[snafu(display("HTTP protocol error: {}", reason))]
-    Http { reason: String },
-
-    #[snafu(display("error parsing {} parameter: {}", param_type, reason))]
-    InvalidParam { param_type: String, reason: String },
-
-    #[snafu(display("unexpected tag in TaggedBase64: {} (expected {})", actual, expected))]
-    TagMismatch { actual: String, expected: String },
-}
+pub use disco_types::request::*;
 
 /// Parameters passed to a route handler.
 ///
@@ -548,17 +508,6 @@ impl RequestParamValue {
                 },
             )
     }
-}
-
-#[derive(
-    Clone, Copy, Debug, EnumString, strum_macros::Display, Deserialize, Serialize, PartialEq, Eq,
-)]
-pub enum RequestParamType {
-    Boolean,
-    Hexadecimal,
-    Integer,
-    TaggedBase64,
-    Literal,
 }
 
 #[derive(Clone, Debug)]
