+++ types/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "disco-types"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+license-file.workspace = true
+
+[features]
+default = ["http-types"]
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+config.workspace = true
+futures.workspace = true
+http.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+snafu.workspace = true
+strum.workspace = true
+strum_macros.workspace = true
+
+http-types = { workspace = true, optional = true }
+
+[dev-dependencies]
+vbs.workspace = true
