+++ types/README.md
@@ -1,0 +1,4 @@
+# disco-types
+
+Shared type definitions used by the [tide-disco](../tide-disco/README.md) web server and
+[surf-disco](../surf-disco/README.md) client.
