+++ types/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod error;
+pub mod request;
+pub mod status;
